+++ plugins/tron/common.go
@@ -1,0 +1,56 @@
+package main
+
+const tronPluginName = "tron-hdwallet-plugin"
+
+// DO NOT EDIT THESE VARIABLES DIRECTLY. These are build-time constants
+// DO NOT USE THESE VARIABLES IN APPLICATION CODE. USE commonConfig.NewLdFlagsManager SERVICE-COMPONENT INSTEAD OF IT
+var (
+	// ReleaseTag - release tag in TAG.SHORT_COMMIT_ID.BUILD_NUMBER.
+	// DO NOT EDIT THIS VARIABLE DIRECTLY. These are build-time constants
+	// DO NOT USE THESE VARIABLES IN APPLICATION CODE
+	ReleaseTag = "v0.0.0-00000000-100500"
+
+	// CommitID - latest commit id.
+	// DO NOT EDIT THIS VARIABLE DIRECTLY. These are build-time constants
+	// DO NOT USE THESE VARIABLES IN APPLICATION CODE
+	CommitID = "0000000000000000000000000000000000000000"
+
+	// ShortCommitID - first 12 characters from CommitID.
+	// DO NOT EDIT THIS VARIABLE DIRECTLY. These are build-time constants
+	// DO NOT USE THESE VARIABLES IN APPLICATION CODE
+	ShortCommitID = "00000000"
+
+	// BuildNumber - ci/cd build number for BuildNumber
+	// DO NOT EDIT THIS VARIABLE DIRECTLY. These are build-time constants
+	// DO NOT USE THESE VARIABLES IN APPLICATION CODE
+	BuildNumber string = "100500"
+
+	// BuildDateTS - ci/cd build date in time stamp
+	// DO NOT EDIT THIS VARIABLE DIRECTLY. These are build-time constants
+	// DO NOT USE THESE VARIABLES IN APPLICATION CODE
+	BuildDateTS string = "1713280105"
+)
+
+func GetPluginName() string {
+	return tronPluginName
+}
+
+func GetPluginReleaseTag() string {
+	return ReleaseTag
+}
+
+func GetPluginCommitID() string {
+	return CommitID
+}
+
+func GetPluginShortCommitID() string {
+	return ShortCommitID
+}
+
+func GetPluginBuildNumber() string {
+	return BuildNumber
+}
+
+func GetPluginBuildDateTS() string {
+	return BuildDateTS
+}